+++ tools/idea-plugin/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Migrate to IDEA code highlighting for tooltip snippets
+
 ## 1.3.0 - 2026-02-28
 
 ### Added